frontend/requirements.txt
websockets>=10.0
playsound>=1.3.0
python-dotenv>=0.19.2
